test: consolidate hello_crypto tests
